requirements.txt
fastapi
uvicorn
llama-index
llama-index-vector-stores-chroma
sentence-transformers
chromadb
python-dotenv
openai
